requirements.txt
aiofiles==22.1.0
aiohttp==3.8.5
aiosignal==1.3.1
aiosqlite==0.18.0
altgraph==0.17.4
anyio==3.5.0
appdirs==1.4.4
appnope==0.1.2
argon2-cffi==21.3.0
argon2-cffi-bindings==21.2.0
asttokens==2.0.5
async-timeout==4.0.3
attrs==22.1.0
Babel==2.11.0
backcall==0.2.0
beautifulsoup4==4.12.2
bleach==4.1.0
blinker==1.5
Bottleneck==1.3.5
brotlipy==0.7.0
certifi==2023.11.17
cffi==1.15.1
charset-normalizer==2.0.4
click==8.1.7
colorama==0.4.6
comm==0.1.2
contourpy==1.0.5
cryptography==41.0.2
cx-Freeze==6.15.10
cycler==0.11.0
debugpy==1.6.7
decorator==5.1.1
defusedxml==0.7.1
docker-pycreds==0.4.0
entrypoints==0.4
et-xmlfile==1.1.0
executing==0.8.3
fastjsonschema==2.16.2
fonttools==4.25.0
frozenlist==1.3.3
gitdb==4.0.10
GitPython==3.1.36
h11==0.14.0
h2==4.1.0
hpack==4.0.0
hypercorn==0.14.4
hyperframe==6.0.1
idna==3.4
ipykernel==6.25.0
ipython==8.12.0
ipython-genutils==0.2.0
itsdangerous==2.1.2
jedi==0.18.1
Jinja2==3.1.2
joblib==1.3.2
json5==0.9.6
jsonschema==4.17.3
jupyter_client==7.4.9
jupyter_core==5.3.0
jupyter-events==0.6.3
jupyter-server==1.23.4
jupyter_server_fileid==0.9.0
jupyter_server_ydoc==0.8.0
jupyter-ydoc==0.2.4
jupyterlab==3.6.3
jupyterlab-pygments==0.1.2
jupyterlab_server==2.22.0
kiwisolver==1.4.4
lxml==4.9.2
macholib==1.16.3
MarkupSafe==2.1.1
matplotlib==3.6.2
matplotlib-inline==0.1.6
mistune==0.8.4
modulegraph==0.19.6
multidict==6.0.2
munkres==1.1.4
nbclassic==0.5.5
nbclient==0.5.13
nbconvert==6.5.4
nbformat==5.7.0
nest-asyncio==1.5.6
notebook==6.5.4
notebook_shim==0.2.2
numexpr==2.8.4
numpy==1.25.0
openai==0.28.0
openpyxl==3.0.9
packaging==23.0
pandas==1.5.3
pandas-stubs==2.0.3.230814
pandocfilters==1.5.0
parso==0.8.3
pathtools==0.1.2
pexpect==4.8.0
pickleshare==0.7.5
Pillow==9.4.0
pip==23.2.1
platformdirs==2.5.2
plotly==5.17.0
ply==3.11
priority==2.0.0
prometheus-client==0.14.1
prompt-toolkit==3.0.36
protobuf==3.20.3
psutil==5.9.0
ptyprocess==0.7.0
pure-eval==0.2.2
py2app==0.28.6
pycparser==2.21
Pygments==2.15.1
pyOpenSSL==23.2.0
pyparsing==3.1.1
PyQt5-sip==12.11.0
pyrsistent==0.18.0
PySocks==1.7.1
python-dateutil==2.8.2
python-docx==0.8.11
python-dotenv==1.0.0
python-json-logger==2.0.7
pytz==2022.7
PyYAML==6.0
pyzmq==23.2.0
qtconsole==5.4.2
QtPy==2.2.0
quart==0.18.4
quart-cors==0.7.0
requests==2.31.0
rfc3339-validator==0.1.4
rfc3986-validator==0.1.1
scikit-learn==1.2.2
scipy==1.11.1
Send2Trash==1.8.0
sentry-sdk==1.31.0
setproctitle==1.2.2
setuptools==68.0.0
sip==6.6.2
six==1.16.0
smmap==3.0.5
sniffio==1.2.0
soupsieve==2.4
stack-data==0.2.0
tenacity==8.2.3
terminado==0.17.1
threadpoolctl==3.2.0
tinycss2==1.2.1
toml==0.10.2
tornado==6.3.2
tqdm==4.66.1
traitlets==5.7.1
types-pytz==2023.3.1.0
typing_extensions==4.7.1
urllib3==1.26.16
wandb==0.15.10
wcwidth==0.2.5
webencodings==0.5.1
websocket-client==0.58.0
Werkzeug==2.3.7
wheel==0.38.4
wsproto==1.2.0
y-py==0.5.9
yarl==1.8.1
ypy-websocket==0.8.2
